+++ src/fides/api/ops/api/v1/endpoints/privacy_experience_endpoints.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from fastapi import Depends, HTTPException, Security
+from fastapi import Depends, HTTPException, Request, Security
 from fastapi_pagination import Page, Params
 from fastapi_pagination import paginate as fastapi_paginate
 from fastapi_pagination.bases import AbstractPage
@@ -13,9 +13,11 @@ from fides.api.ops.api.v1 import scope_registry
 from fides.api.ops.api.v1 import urn_registry as urls
 from fides.api.ops.models.privacy_experience import ComponentType, PrivacyExperience
 from fides.api.ops.models.privacy_notice import PrivacyNotice, PrivacyNoticeRegion
+from fides.api.ops.models.privacy_request import ProvidedIdentity
 from fides.api.ops.oauth.utils import verify_oauth_client
 from fides.api.ops.schemas.privacy_experience import PrivacyExperienceResponse
 from fides.api.ops.util.api_router import APIRouter
+from fides.api.ops.util.consent_util import get_fides_user_device_id_provided_identity
 
 router = APIRouter(tags=["Privacy Experience"], prefix=urls.V1_URL_PREFIX)
 
@@ -54,6 +56,8 @@ def privacy_experience_list(
     component: Optional[ComponentType] = None,
     has_notices: Optional[bool] = None,
     has_config: Optional[bool] = None,
+    fides_user_device_id: Optional[str] = None,
+    request: Request,
 ) -> AbstractPage[PrivacyExperience]:
     """
     Returns a list of PrivacyExperience records for individual regions with
@@ -61,8 +65,17 @@ def privacy_experience_list(
 
     'show_disabled' query params are passed along to further filter
     notices as well.
+
+    'fides_user_device_id' query param will stash the current preferences of the given user
+    alongside each notice where applicable.
     """
     logger.info("Finding all Privacy Experiences with pagination params '{}'", params)
+    fides_user_provided_identity: Optional[ProvidedIdentity] = None
+    if fides_user_device_id:
+        fides_user_provided_identity = get_fides_user_device_id_provided_identity(
+            db=db, fides_user_device_id=fides_user_device_id
+        )
+
     experience_query = db.query(PrivacyExperience)
 
     if show_disabled is False:
@@ -90,7 +103,9 @@ def privacy_experience_list(
     ):
         privacy_notices: List[
             PrivacyNotice
-        ] = privacy_experience.get_related_privacy_notices(db, show_disabled)
+        ] = privacy_experience.get_related_privacy_notices(
+            db, show_disabled, fides_user_provided_identity
+        )
         privacy_experience.privacy_notices = privacy_notices
         if not (has_notices and not privacy_notices):
             results.append(privacy_experience)
@@ -111,11 +126,16 @@ def privacy_experience_detail(
     db: Session = Depends(deps.get_db),
     privacy_experience_id: str,
     show_disabled: Optional[bool] = True,
+    fides_user_device_id: Optional[str] = None,
+    request: Request,
 ) -> PrivacyExperience:
     """
     Return a privacy experience for a given region with relevant notices embedded.
 
     show_disabled query params are passed onto optionally filter the embedded notices.
+
+    'fides_user_device_id' query param will stash the current preferences of the given user
+    alongside each notice where applicable.
     """
     logger.info("Fetching privacy experience with id '{}'.", privacy_experience_id)
     experience: PrivacyExperience = get_privacy_experience_or_error(
@@ -128,8 +148,14 @@ def privacy_experience_detail(
             detail=f"Query param show_disabled=False not applicable for disabled privacy experience {privacy_experience_id}.",
         )
 
+    fides_user_provided_identity: Optional[ProvidedIdentity] = None
+    if fides_user_device_id:
+        fides_user_provided_identity = get_fides_user_device_id_provided_identity(
+            db=db, fides_user_device_id=fides_user_device_id
+        )
+
     # Temporarily stash the privacy notices on the experience for display
     experience.privacy_notices = experience.get_related_privacy_notices(
-        db, show_disabled
+        db, show_disabled, fides_user_provided_identity
     )
     return experience